% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant_coordinates.R
\docType{data}
\name{ant_coordinates}
\alias{ant_coordinates}
\title{Antenna coordinates.}
\format{
A data frame with xxx rows and yy variables:
\describe{
  \item{antenna}{antenna number}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{z}{z coordinate}
  ...
}
}
\usage{
ant_coordinates
}
\description{
A dataset containing the locations of pond entrance antennae.
}
\keyword{datasets}
