% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_social_network.R
\name{pr_make_network}
\alias{pr_make_network}
\title{Construct social networks from reads data.}
\usage{
pr_make_network(block_df, block_ref_df, time_bin = "5 seconds")
}
\arguments{
\item{block_df}{A data frame containing the reads from an experimental block.
This data frame must have at least the following columns:
`time`, `antenna`, `id`.}

\item{block_ref_df}{A data frame containing a reference list with all
individuals present in the experimental block.}

\item{time_bin}{A time bin that can be understood by `lubridate::round_date`,
such as "5 seconds", "1 minute", or similar.}
}
\value{
A `tidygraph` network object.
}
\description{
Function to return a `tidygraph` network object from fish
detected within a `time_bin` at one of multiple readers.
}
\examples{

data(block_df)
data(block_ref_df)

network = pr_make_network(block_df, block_ref_df, time_bin = "5 seconds")

}
